Add bitrate, start time and chapters video metadata to Inspector - Fix backend BigInt conversion incorrectly using i32 instead of u32 - Change FFmpegFormatContext/FFmpegMetaData bit_rate to i64 - Rename byteSize to humanizeSize - Expand humanizeSize logic to allow handling bits and Binary units - Move capitalize to @sd/client utils

+++ crates/ffmpeg/src/model.rs
@@ -7,7 +7,7 @@ pub struct FFmpegMediaData {
 	pub formats: Vec<String>,
 	pub duration: Option<i64>,
 	pub start_time: Option<i64>,
-	pub bit_rate: i32,
+	pub bit_rate: i64,
 	pub chapters: Vec<FFmpegChapter>,
 	pub programs: Vec<FFmpegProgram>,
 	pub metadata: FFmpegMetadata,
